requirements.txt
timm==0.6.12
torchvision==0.14.1
torchmetrics==0.11.4
pytorch-lightning==1.7.3
matplotlib
opencv-python
tqdm
imageio
path
scipy
configargparse
kornia
